+++ overlay/g350/roms/ports/cozos/overlay.py
@@ -9,7 +9,7 @@ import re
 import subprocess
 import sys
 
-VERSION = '0.5.1'
+VERSION = '0.5.2'
 ROOT = Path(os.environ.get('COZOS_ROOT', '/'))
 STATE = ROOT / 'userdata/system/cozos'